Split boot.rs; delete Darwin ensure_shim_dylib_aliases
Move managed boot into boot/{mod,spawn,guest,unix}.rs with re-exports.
Delete the fail-open Darwin dylib copier. Drop unused spawn_shim vm_id.
